runtime: Reimplement (call/cc) using setjmp
This is a reimpementation of (call/cc) using stack copying and
setjmp/jongjmp. This saves a "continuation base" which is the stack
pointer of the launchWorld() function. Then, when (call/cc) is invoked
the stack from the continuation base is copied on to the heap and it's
GC-related data structures are relocated to point to the new addresses.

When the garbage collector encounters an escape procedure it will
extract its continuation and then visit its relocated shadow stack,
strong refs and weak refs. This code is shared with the normal GC path.

Upon invoking an escape procedure the stack is copied back in to place
and the GC data structures are "delocated" back to their original
address.

(call/cc) works in the cases it did previously in addition to now
supporting multiple invokations and calling "down" the callstack.
However, the interaction with dynamic states is completely broken and
needs to be fixed in the following commits.