stdlib: Implement (assv), (assq) and (assoc)
These are implemented very similiarily to the (member) family of
procedures.

An initial implementation in Scheme revealed a weakness with the current
implementation of recursive procedures: they're always considered to
introduce a GC barrier even if they never allocate memory or call
worldfunctions. This is because the procedure is conservatively assumed
to be a world function during initial planning and when the planned
function is analysed it's determined to use the world pointer during its
recursive call. This should be improved in the future.