Initial vector type support
Vector types are actually composed of three different types:

- Uniform vectors are constructed using (Vectorof) and consist of a
  vector containing all of the same element. This type is unstable and
  degrades to (Vectorof <any>) for storage locations.

- Specific vectors are constructed using (Vector) and denote of a vector
  of known length and a known type for each element. This type is also
  unstable and degrades to a (Vector <any> ...) with the same number of
  elements as the original vector. This is useful as a vector's length
  can be used to optimise in-bounds vector access in the future

- The exisiting vector SchemeTypeAtom is the preferred alias for
  (Vectorof <any>) but behaves identically as far as type checking

These types support full type recursion in the same way that pairs due.
This validates that the recursive type infrastructure is general enough
to support other types.

Due to their unstable nature these types will generally decrease
performance as the vector needs to be explicitly checked for a type. It
is useful for type safety and documentation, however.

Type checking of uniform vectors introduced a new ps.ForAll plan step
which checks that a predicate holds true for an entire array. This
wasn't easy to accomplish using normal Scheme tail recursion as vector
types can't be easily "sliced" in to a smaller vector and there's no way
to tell if the end of the vector has been reached without explicitly
passing an index. There would also be no way to track the generated
vector data "pseudo-predicates" with the exisiting type check data
structures.