compiler: SImplify unconditional ForAll IR
(make-vector) now uses ForAll without any loop body condition to allow
early breaks. The IR it generated worked but was a bit silly. Instead
split GenForAll to its own function and have it detect and simplify
loops without an exit condition

With optimisation this should result in the same assembler but it's
easier to read the IR this way.