compiler: Better planning of (apply)
(apply) without a known argument list would previously unconditionally
bounce through a procedure trampoline even if the exact signature of the
calling function was known. This was suboptimial but tolerable before
multiple values but now procedure trampolines need to allocate to place
single values in an arbitrary values result list.

To fix this factor out the trampoline planning code in to a new
PlanInvokeApply.withArgumentList. This will use type information from
the passed list value to remove type checks where possible. It will also
return the ResultValues from the function directly instead of placing it
in an ArbitraryValues list unconditonally.