stdlib: Add implementation of (guard)
This uses the (guard) and (guard-aux) macros from R7RS + errata as a
basis. However, R7RS's (guard) uses two (call/cc)s, (call-with-values)
and (with-exception-handler). This is very inefficient for our Scheme
implementation.

Instead re-implement the core (guard) macro in the C++ stdlib as
_lliby_guard_kernel. If no exception is thrown the guard kernel is
nearly zero overhead and has no GC allocations.

Keep the exception matching logic in (guard-aux) identical to R7RS as
that's strongly suited to a Scheme implementation.