compiler: Test for typed recursive capturing procs
This test consists of a typed procedure that takes a value and the
returns typed recursive procedure returning that value. This also adds a
new (assert-raises) test helper to ensure an expression raises an
exception

This isn't testing anything in particular; it's just ensuring that a
number of tricky features interact properly in the same procedure.