compiler: Fix procedures placed in pairs
When placing a procedure in a pair we would convert to the procedure to
<any> which would give it a trampoline with the TopProcedureCell
signature. However, we were storing the original type of the procedure
inside the iv.KnownPairCellValue which means we would try to use the
original signature when invoking the cell.

Add vt.ValueType.replaceProcedureType to let us rewrite to Scheme type
to indicate that we converted the pair to the TopProcedureType. Change
iv.KnownPairCellValue to track the actual stored procedure type.

This is preferrable to preserving the procedure type as that would
require a mechanism to convert pairs of different procedure types
between each other which would be complex and fragile. This way all
pairs have the top procedure type and can freely convert between each
other.

Additionally, add a fix for duplicate error message strings when
creating multiple adapter procedures with different arities. This was
discovered while writing one of the functional tests for the above
issue.