compiler: Default to lists being capturable
Due to a typo the planner was assuming all lists were non-capturable.
For dialects with immutable pairs this was harmless but for R7RS this
meant it assumes the types of lists could never change. This meant
mutating lists could easily break the type system.

Default to assuming lists aren't capturable and add a test.