compiler: Fix multiple conversions of same proc
Whenever a KnownProc cell was converted to a new type we were
overwritting the entry point with the required adapter procedure.
However. However, if a procedure was converted to multiple types the
previous conversions would have their entry point overwritten.

Fix this by treating KnownProc procedure conversions like converssions
of unknown cell values: create a new adapter procedure cell that points
to the original closure.

This is less efficient if the original closure won't be directly used
but this fixes a real bug and can be optimised later