compiler: Restructure libraries
(llambda stdlib) both defined (case-lambda) and used it internally in
its definitions. This is in contrast to R7RS where (case-lambda) was a
separate library.

Now that (case-lambda) is a primitive we can structure our libraries
more closely to how they are in R7RS. Namely, everything that doesn't
add exports over R7RS is implemented directly in the "scheme" directory
while those adding exports are defined in the "llambda" directory with a
fascade library in the "scheme" directory if applicable