compiler: Avoid GC flush if branch terminates
If one or both of our CondBranch branches unconditionally terminates we
don't need to merge our GC state once the branch completes. This means
we don't need to flush our GC state before the branch.

This is useful for tail recursive procedures that take GC managed
arguments. Before, we would create a shadow stack entry and root the
argument before the conditional branch containing the tail call even if
there was no other GC usage in the procedure.