compiler: Fix mutually recursive type pred procs
This was encountered as a sporadic failure planning the type predicate
for (Listof <any>). Due to another issue the order which union member
types are tested in a type check is undefined. Sometimes (Listof <any>)
would be planned as:

- Check for <empty-list>
- Check for (Pairof <any> (Listof <any>))

The Pairof test would attempt to perform another test for (Listof <any>)
but it wasn't aware that test was already being planned. The result was
the generation of an unbounded number of type predicates until the Scala
stack was exhausted.

Instead of building the type predicate system around
KnownTypePredicateProc move the equivalent functionality in to a
dedicated SymbolForTypePredicateProc function. This function will note
the allocated symbol for a type predicate proc before it starts planning
that proc which allows an arbitrary number of recursive type predicate
procedures to be planned at once. The hack in PlanTypeCheck to deal with
self-recursive type predicates has been removed as the more general fix
now covers it.

For some reason this is difficult to trigger during functional tests.
There is some upcoming tests for other features that trigger this about
50% of the time that will hopefully be sufficient to catch this in the
future.