compiler: Prefer unboxed arguments for lambdas
The justification behind accepting boxed types was that if the caller
had already allocated a box for the value we wanted to re-use it instead
of causing another allocation. This was chosen over the downside of
potentially having to force the caller to allocate if it didn't already
have a box.

A few things have changed since that decision:

- The type system and explicit typing is much more sophisticated.
  Presumably performance sensitive code will use type annotations that
  avoid having to explicitly box values in most situations

- We support more inline planning for operations such as integer math.
  These typically operate on and return unboxed values

- Because of #1 and #2 tail recursive procedures using e.g. integer
  accumulators or counter need to box a value simply to tail call
  themselves. Now that somewhat efficient tail calls are supported this
  is a barrier to implement stdlib behaviour using tail recursion