compiler: Fix integer math at -O 2
We were passing nuw (no unsigned wrap) to our integer arithmetic
instructions. Because we deal with signed values (LLVM has no concept of
signedness in its type system) passing nuw is incorrect and led to
undefined results for signed values that did not wrap.

Add a test that initially showed this issue. This is tricky to test
because most trivial tests will get collapsed in to a compile time
    constant.