Include file and line in runtime error messages
This extends signalError and its codegen entry point _lliby_signal_error
to include "path" and "lineNumber" arguments. These are appended to the
error message at runtime based on the source location of the plan step
that is generating the error. This reuses the same location tracking
that debug info generation uses so it's fairly simple to implement.

The file and line number are passed separately instead of being baked in
to the message string to allow the error messages and path strings to be
reused across multiple messages. In the future that information could
also be explicitly included in the error object.

Closes #45