compiler: Planning for many-valued number compare
The existing planner for (=), (<), etc. only supports planning with two
operands. This was under the assumption that planning with many operands
is relatively rare and not worth the complexity. However, (<= lower
value upper) is an idiomatic way of performing a range check. It turns
out that a type annotated version of the R7RS life example spends more
than half its time in lliby_numeric_lte checking vector bounds.

Implement by doing a sliding comparison on each pair of values and then
using ps.CondBranch to merge their results together. This generates the
same plan for two operands as it did previously and has been confirmed
to produce sane assembler on x86-64 in the many-operands case.

This also adds a fairly awkward functional test for this code path.
It's difficult to trigger with more straightforward tests as they will
have their results statically evaluated.