compiler: Support for (/) planning
This supports planning of (/) with static values or dynamic values of
known types. This works similarily to how (+), (*) etc are implemented
except it always uses floating point instructions even if all the
operands are integers. This is because unlike the other arithmetic
procedures floating point values can be produced by two integer
operands.

This also adds functional tests for typed procedures wrapping (*), (-)
and (/). This ensures that their native instruction generation code
paths are exercised properly.