compiler: Static planning for (cons) and (append)
The primary motivation behind this is to allow optimised planning of
(apply) in more situations.

When performing tail calls to procedures with rest arguments (apply)
might be needed to pass the rest argument. Our fallback (apply) path in
the stdlib doesn't guarantee tail recursion so it needs to be avoided
whenever possible.