compiler: Better planning for double equivalence
Normally (eqv?) on flonums can't be trivially planned because in Llambda
(eqv? +nan.0 +nan.0) is true. There is no LLVM condition that
corresponds to that test. However, if one of the flonum values is known
then we can either compare the float values or check for NaN

This adds a new ps.FloatIsNaN step that checks if a floating point value
is NaN. This works by checking if the value is not equal to itself which
is only tue for NaN.