compiler: Add tests for +nan.0 comparisons
This ensures that +nan.0 is unordered in various situations. This
already worked correctly but it's a tricky area especially with respect
to optimisations

This also introduces (typed-dynamic) as a macro that wraps:
(cast (typeless-cell value) <type>)

This is a fairly frequent pattern to erase a value's identity but
constraining it to a specific type. This is useful to test the type
system or certain optimisations (in this case the native floating point
generation).