compiler: Planner support for integer division
This adds planning for (truncate/), (truncate-quotient) and
(truncate-remainder). If both the denominator and numerator are static
they will be evaluated at compile time. If only the denominator is
static then native instructions will be generated to calculate at
runtime.

Dynamic denominators are unsupported to avoid having to create inline
checks for division by zero.