compiler: Fix recursives capturing non-recursives
When the self-recursive lambda optimisation was introduced we "hid" all
the values from the same binding from PlanLambda so it wouldn't consider
the self value captured. This had the side-effect of making all other
values in the same binding invisible to the lambda. We were also
returning the initial state in to the fold which meant that all forward
declared recursive values were forgotten when the first self-recursive
value is encountered.

Fix by explicitly excluding the recursive self value in FindClosedVars
and tracking the current state properly in PlanBind