runtime: Clean up StringCell::unicodeChars
This was originally implemented before strings had a known character
length. As the length was unknown special logic was required to detect
reading past the end of the string and dealing with end == -1 meaning
end of string.

This uses the character length of the string to clean up the logic and
allow us to fail early in more cases.