Support exact (/)
If exact integers that divide in to each other without a remainder are
passed to (/) we should return an exact result. For example, the
following should be true:

(eq? 5 (/ 20 10 2))

Previously we were always returning an inexact value for (/) which isn't
ideal from a consistency or Scheme compatibility perspective.

This is slower as the result type of (/) between two dynamic integers is
unknown and requires a library call to calculate.  However, performance
sensitive code should be using explicit number types.