Split out arithmetic implementation
NumberProcPlanner.scala in the compiler and number.cpp in the runtime
were reaching an unwieldy size. Separate out the pure arithmetic
operations to ArithmeticProcPlanner.scala and arithmetic.cpp while
leaving the comparison etc. operations in the original files.

This also creates an inexact.cpp for the (finite?) and (infinite?)
implementation. This reflects the fact they're placed in a different
Scheme library from the other runtime functions.