compiler: Prefer compact return types from lambdas
When planning typed recursive lambdas prefer compact (e.g. unboxed)
values. This avoids having to box values simply to return them. For
non-recursive typed lambdas the the preferred return type will still be
used.

This is a similar tradeoff to a accepting compact arguments - this is
only a win if the procedure computes using unboxed arguments.