compiler: Have adapted procs use compact types
When building a procedure signature from a Scheme procedure type prefer
unboxed values. This matches what we do with planning lambdas.

This was motivated by the (string-map) where the majority of its mapping
functions will not want to box and unbox character values. This allows
them to take simple 32bit integers representing the character's code
point.