compiler: Make all values convertable to <unit>
The motivation behind this is to allow procedures that return values to
be passed as procedure types returning <unit>. This allows the
procedure's value to be safely discarded. This complicates the planning
for (eqv?) with #!unit but that is such a corner case that support can
be removed.

This is analogous to how any expression in Scala can be cast to Unit to
discard the value