runtime: Remove confusing VectorCell::fromElements
There is an unused VectorCell::fromElements() builder that attempts to
take ownership of the std::vector data. Since the pointer is actually
owned by the passed std::vector this is incorrect. This builder is
unused so just remove it completely

Also add a Doxygen comment for the remaining ::fromElements() to be
clear that it expects to take ownership