compiler: Proper (*for-each) support
Previously (for-each) and (vector-for-each) were implemented as aliases
for (map) and (vector-map) respectively. This had a number of downsides:

- Lambdas returning multiple values could not be passed to (*for-each)
  as the value was being used to build a temporary result. This problem
  was much worse for a potential (string-for-each) implementation as it
  would require the lambda to return a char

- It's possible to accidentally depending on the return value from
  (for-each) and (map) which is unspecified by R7RS

- We waste resources building and discarding an unused result value

Fix this by moving the mapping implementation to map.cpp and abstracting
away the mapping core logic for each type. New entry points are defined
for the (*for-each) functions that take a lambda returning <void> and
don't return a result value.

This isn't completely optimal for (*for-each) as we still build and GC
root an output vector while iterating over the input collection.
However, for lists and strings we can avoid the (potentially expensive)
construction of the final value.