runtime: Don't buffer string likes and write
When writing symbols and strings we were writing to a temporary
ostringstream until we determined if the string-like needed quotes. This
has a number of disadvantages:

- Constructing an ostringstream is expensive (at least on libstdc++) due
  to locale setup

- This requires allocating additional memory for the output string with
  no clean way to reserve a buffer. This can lead to many allocations
  for large strings.

- The writing code mixed rendering logic with quote tracking logic. This
  makes it difficult to follow

Instead first do a pass on symbols to determine if they need a quote.
This can abort on the first quote-requiring character and can be skipped
completely for strings which are always quoted. Then, write the output
directly to the writer's output stream

On a modified version of generate-cxr-library.scm that outputs ~50K
lines of Scheme this more than halved its runtime. This should be even
more dramatic for strings which can skip the quote detection completely.