runtime: Optimise UTF-8 encoding
Currently fromUnicodeChars() and toConvertedString() use
encodeUtf8Char() to repeatedly insert a temporary vector to the end of
an output vector. This requires a temporary allocation which is very
slow.

Instead, introduce appendUtf8Char() which will append the encoded
character to any output iterator. This allows us to use
std::back_inserter to append to the UTF-8 output vector without using a
temporary. utf8::encodeUtf8Char() is reimplemented as a simple wrapper
on top of appendUtf8Char() for use cases that only need a single
character.

This results in an approximately 15% speedup in the generate-cxr-library
benchmark which calls ::fromUnicodeChars() via (list->string)