compiler: Don't box when phi'ing consts and native
This makes phi'ed values between conditional branches use the unboxed
value when phiing constant values with native values or other constants.
Previously they would use a boxed value which would cause boxing if the
other values was a native value.

This is a less clear advantage when phi'ing two constant values. If the
value later needed to be boxed we'll needlessly create a GC allocated
cell instead of phi'ing two constant values. However, if that phi'ed
constant value is returned from a procedure or phi'ed with another
non-constant value it will no longer be known to be a constant and will
require GC rooting.

- Add a new iv.UnboxedValue trait that indicates values that can be
  immediately represented by a native value

- Have all NativeValue classes implement UnboxedValue

- Have Constant{Flonum,ExactInteger,Char,Boolean}Value implement
  UnboxedValue

- Change PlanResultValuesPhi to look for the UnboxedValue trait instead
  of NativeValue