compiler: Ensure lambda arguments names are unique
With macros its possible for a lambda to have the an argument with the
same name in different scopes. Currently this will cause us to generate
invalid IR due to duplicate argument names. Use SourceNameUniquer which
will assign duplicate names distinct suffixes.

This is difficult to trigger due to a macro expander bug but it will be
indirectly tested by a following commit.