compiler: Rescope all symbols in expanded template
This is to support the (let-values) macro which recursively expands
itself introducing multiple arguments named "x" to an expanded lambda.
This currently fails as we're not implementing the following behaviour
mentioned in R7RS:

"If a macro transformer inserts a binding for an identifier (variable or
keyword), the identifier will in effect be renamed throughout its scope
to avoid conflicts with other identifiers"

This can be implemented by our frontend by assigning all symbols fresh
scopes that contain their existing scope as a parent. They will then be
considered distinct symbols even if they were expanded from the exact
same symbol in the source.

This is difficult to directly test but will be validated by the
(let-values) test in an upcoming commit.