Add <eof-object> as distinct type
(read) needs to return an eof-object when it encounters an end-of-file.
This means it cannot be any datum that can be read using (read).
Currently <eof-object> is an alias for <unit> which can be represented
as #!unit

For the same reason there is no reader form for <eof-object>. It can
only be obtained using (eof-object).