runtime: Support for error object categories
This will be used to implement (read-error?) and (file-error?).
Currently there is only one error category of "Default" that is used as
a catch all for the current error types. This functionality is also not
exposed through _lliby_signal_error - all errors produced by codegen are
in the Default category.

This could be extended in the future to allow for something similar to
R6RS's condition types if it's deemed useful.

This also moved the line/column number formatting logic in to
_lliby_signal_error. There is no point exposing that functionality to
C++ stdlib code as it has no source location information. This was done
to keep the signature for signalError() under control.