stdlib: Implement (numerator), (denominator)
For exact integers this is trivial. We use the value as the numerator
and 1 as the denominator

For flonums this is much trickier. We use frexp() to split the float in
to a significand and exponent. We then reduce the resulting fraction to
lowest terms.

This produces a numerator and denominator that exactly capture the
original flonum. However, this is subject to the issues of floating
point representations. For example, 0.1 can't be represented exactly in
binary floating point and its closest floating point value produces a
fraction with very large terms.