runtime: Implement (rationalize)
This is a pretty dumb algorithm. It depends on the fact that flonums can
only exactly represent fractions with powers of two as the denominator.
It reuses the (numerator)/(denominator) infrastructure to split the
number. It then repeatedly divides the numerator and denominator by two
while tracking the last candidate that passed the error check.