runtime: Fix (string-map) crash on x86-32
Apparently in the x86-32 ABI returning a class/struct with a single
member is not the same as returning the member type. This caused a crash
(string-map) when it tried to invoke the mapping function with the wrong
signature.

Instead introduce UnicodeChar::CodePoint as an alias for std::int32_t
and define the mapper function to return that.