runtime: Workaround read-bytevector! bug on x86-32
Under the following circumstances we may be unable to catch an exception
thrown by a native function:

- The target platform is x86-32

- tailcallopt is enabled with llc and an explicit tail call is placed in
  the right place in the call chain.

- The stack is large enough. Removing arguments from the native function
  or optimising out the (let)'s SEL (as in -O2) causes the catch to
  succeed again

This has the hallmarks of a possible LLVM bug. x86-32 uses a different
calling convention for tail calls and it's not as well tested as the
default calling convention. It's unsurprising it could have bad
interactions with exception handling.

To work around this we can move the native-read-bytevector! definition
from a (let) to a top-level (define). This reduces the depth of our call
stack at -O 0 and allows the exception to be caught again. It also
generates significantly better code on all platforms at -O 0