runtime: Fix huge heap waste on _LLIBY_ALWAYS_GC
The _LLIBY_ALWAYS_GC logic to force a GC cycle was happening at a very
low level. Whenever a new memory segment was allocated it was
immediately sealed by setting its "next" pointer to its "end" pointer.
The intent was the force program to re-enter the runtime for the next
allocation and force another GC cycle.

However, this had the side effect of forcing every cell being moved to
the new heap during garbage collection to allocate a new segment. This
was compounded by the fact that segments allocated from the same heap
use successively larger allocations up to 1MiB. The result was nearly
every cell moved during garbage collection will use 1MiB of virtual
memory and 4KiB of physical memory. This is pathologically wasteful and
quickly exhausts virtual memory on 32bit platforms.

Instead of sealing a segment from allocations after it's first allocated
instead seal it before returning control back to the program. This
allows the segments to be completely used during garbage collection
while still forcing a GC on every allocation.