stdlib: Implement (char-ready?)
This is built on top of the same bytesAvailable() function as
(u8-ready?). It's fairly complex as it need to ensure an entire UTF-8
byte sequence is ready. It does this by reading the header byte then
attempting to read the reset of the sequence. After detecting if the
read would block we then put the read bytes back.