stdlib: Add (llambda complex-stub)
This has the same interface as (scheme complex). However, as numbers
with imaginary parts are not supported it's of little practical use.
It's not named (scheme complex) to prevent programs doing feature
detection with (cond-expand) from determining the complex numbers are
supported.

This is mainly intended to be re-exported as part of the r5rs library