stdlib: Implement (scheme time)
This makes a very lazy attempt at implementing TAI for (current-second).
It will first try CLOCK_TAI if available and then fall back to
CLOCK_REALTIME with a fixed offset. CLOCK_TAI is Linux-specific and
unexposed by glibc so this won't work without manual hacking.

(current-jiffy) is implemented as CLOCK_MONOTONIC and falls back to
CLOCK_REALTIME if that's not available.

On Mac OS X clock_gettime() is unsupported so gettimeofday() is used
instead. Unfortunately this only have microsecond resolution.