Support for non-recursive datum labels
This adds support for datum labels. This is useful for sharing repeated
data structures in s-expressions.

Unlike R7RS datum labels aren't defined until the datum is fully parsed.
This means it's not possible to define recursive data structures using
labels. It would be significant work to add support for recursive ASTs
in the compiler even if it was supported by the parser.

Closes #35