runtime: Remove StringCell::allocSlackBytes()
StringCell internally stored the number of extra bytes in its
SharedByteArray allocation inside a 16bit integer. This had a number of
issues:

- The logic to track the allocation slack was complex and error-prone.

- More than 32KiB of slack could not be recorded

- When adopting a SharedByteArray from a SymbolCell or BytevectorCell
  the slack would not be known

Instead add a SharedByteArray::capacity() so we can directly calculate
our shared byte array capacity. This has the disadvantage of not being
able to detect slack without a platform specific implementation but it
should be worth the complexity reduction.