compiler: Limit the size of (make-vector)'s type
The report procedure planner for (make-vector) will generate a
SpecificVectorType for the created vector if the length argument is
known. For very large vectors this could result in a type with hundreds
of thousands of elements. When performing type arithmetic etc. on those
types the compiler could effectively livelock or abort due to GC
overhead.

Instead limit the SpecificVectorType size to 1024 in (make-vector) and
fall back to UniformVectorType if that limit is exceeded.