compiler: Better plan for constraining to literals
If a value has its type constrained to a literal type it can be
replaced with a constant of that type. This allows us to discard the
original value which reduces GC pressure and allows for better code
generation.

Previously only LiteralBooleanType supported this and it was broken for
boxed values (i.e. CellValue)

- Introduce IntermediateValue.fromLiteralType that will convert a
  literal type to a constant value representing that literal

- Add support for converting literal symbol types to constant symbols

- Have the relevant .withSchemeType implementations call fromLiteralType
  when literal types are encountered.